feat: add three dots button in published view item
