+++ frontend/appflowy_flutter/lib/workspace/presentation/settings/pages/sites/constants.dart
@@ -1,0 +1,14 @@
+import 'package:intl/intl.dart';
+
+class SettingsPageSitesConstants {
+  static const threeDotsButtonWidth = 26.0;
+
+  static final dateFormat = DateFormat('MMM d, yyyy');
+
+  static final publishPageHeaderTitles = [
+    // todo: i18n
+    'Page',
+    'Published name',
+    'Published date',
+  ];
+}
